moved form into separate component and fixed a bug
